==> code/bitcoin/01_tweets_cleaning.ipynb <==
#### The following notebook includes:
1. Initial analysis of the data
2. Sampling for computation ease
3. Regular Expression Transformation

## Libraries

import re
import pandas as pd
from tqdm import tqdm

File names:

directory = '~/PycharmProjects/tfm_hugopobil'
tweets_fn = f'{directory}/data/bitcoin_tweets.csv'
bitcoin_price_fn = f'{directory}/dataBTC-USD.csv'

## Visualization

We import the dataset which includes the tweets we have downloaded from Internet. These tweets include date ranges from 10th February 2021 until 2 March 2022.

In order to perform all the computation we will reduce the dataset to 1% of total data. Meaning we will analyse and predict using 23470 initial tweets.

# Do not execute, 2 million tweets coming...
tweets = pd.read_csv(tweets_fn, low_memory=False)
print('Shape :', tweets.shape)
print('Tweets DataFrame Initial Date :', tweets.date[0])
print('Tweets DataFrame Final Date :', tweets.date.iloc[-1])
tweets.head()

tweets.columns

tweets.sort_values(by = 'date').tail(20)

### Stratified sampling:

Data shows some imported rows give error, I delete those last 12 rows to avoid getting more errors.

print(type(tweets.date))
tweets = tweets.sort_values(by = 'date')
tweets = tweets[0:len(tweets)-12]

tweets['sample_date'] = tweets['date'].apply(lambda x: x.split(' ', 1)[0])
tweets['sample_date'] = tweets.sample_date.apply(lambda x: pd.to_datetime(x))

tweets_sample = tweets.groupby('sample_date', group_keys=False).apply(lambda x: x.sample(200))

tweets_sample[tweets_sample.sample_date == '2021-02-05'].count()

tweets_sample.shape

## Data Cleaning

The first step when dealing with tweets is to clean the tweet body data. By appliying the 1% reduction to all tweets, we then apply basic regular expression treatment and save the data to CSV.

from cleantext import clean

# clean tweets
tweets_sample = tweets_sample.sort_values(by = 'date')

# Select a 1% random sample of all tweets
dd = tweets_sample.copy()
dd = dd.reset_index()

for i, s in enumerate(tqdm(dd['text'], position=0, leave=True)):
    text = str(dd.loc[i, 'text'])
    text = text.replace('#', '')
    text = text.lower()
    text = re.sub("@[A-Za-z0-9_]+","", text)
    text = re.sub("#[A-Za-z0-9_]+","", text)
    text = re.sub(r"http\S+", "", text)
    text = re.sub(r"www.\S+", "", text)
    text = re.sub(r'[0-9]', '', text)
    text = clean(text, no_emoji=True)
    text = re.sub('[()!?]', ' ', text)
    text = re.sub('\[.*?\]',' ', text)
    text = re.sub("[^a-z0-9]"," ", text)
    text = re.sub('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+', '', text, flags=re.MULTILINE)
    text = re.sub('@\\w+ *', '', text, flags=re.MULTILINE)
    dd.loc[i, 'text'] = text

dd.to_csv(f'{directory}/data/sampled_data/tweets_clean_v2.csv', header=True, encoding='utf-8', index=False)

for i, s in enumerate(tqdm(dd['text'], position=0, leave=True)):
    text = str(dd.loc[i, 'text'])
    text = clean(text, no_emoji=True)
    dd.loc[i, 'text'] = text

dd.to_csv(f'{directory}/data/sampled_data/tweets_clean_v2.csv', header=True, encoding='utf-8', index=False)

# Some examples of tweets processed
dd.text.iloc[0:10]

### End

Save to local the data after processing visualization and cleaning. We will start the next steps using the CSV 'tweets_clean.csv'
