==> code/bitcoin/02_sentiment_analysis_PNN.ipynb <==
## Sentiment Analysis (clean_tweets) - VADER

VADER (Valence Aware Dictionary and Sentiment Reasoner) is a lexicon and rule-based sentiment analysis tool that is specifically attuned to sentiments expressed in social media.

VADER takes into account:
negations and contractions (not good, wasn’t good) Punctuation (good!!!), CAPS, emotes :), emojis Intensificators (very, kind of), acronyms ‘lol’ Scores between -1.0 (negative) and 1.0 (positive)

We will use this sentiment analysis of the tweets to calculate a score that will represent the importance of each tweet.

import pandas as pd
from pysentimiento import create_analyzer

tweets = pd.read_csv('tweets_clean.csv', low_memory=False)

print(tweets.shape)
tweets.head()

### Create a column in the dataframe with sentiment analysis:

Google Trends buscar Bitcoin, tweets a nivel de país
Probar API Google NLP

analyzer = create_analyzer(task='sentiment', lang='en')

# tweets['sentiment'] = analyzer.predict(tweets['text'])
results = []
for ind, row in tweets.iterrows():
    results.append(analyzer.predict(row.text))
    if ind % 10 == 0:
        print(ind)
print('Done')

## Save to local

results_df = pd.DataFrame(results)
results_df.to_csv('results_sentiment_analysis_PNN.csv')
print('Document saved')

Join to dataframe:

results_df = pd.read_csv('results_sentiment_analysis_PNN.csv')

results_df.shape[0] == tweets.shape[0]
# print(results.shape)

tweets['sentiment'] = results_df
tweets['simplified_sentiment'] = tweets["sentiment"].apply(lambda x: x.output)
tweets.simplified_sentiment.value_counts()

tweets

### Save to local

tweets.to_csv('tweets_scores_PNN.csv')
