==> code/bitcoin/04_NLP_Modelling.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
from textblob import TextBlob

from nltk.tokenize import TweetTokenizer
from nltk.stem.wordnet import WordNetLemmatizer

directory = '~/PycharmProjects/tfm_hugopobil'

btc_usd_grouped = pd.read_csv(f'{directory}/data/sampled_data/btc_usd_grouped_v2.csv')
tweets_grouped = pd.read_csv(f'{directory}/data/sampled_data/tweets_grouped_v2.csv')

df_clean = pd.read_csv(f'{directory}/data/sampled_data/tweets_clean_v2.csv')
df_clean.head()

df_clean.dropna(subset=['hashtags'], inplace=True)
df = df_clean[['text']]
df.columns = ['tweets']
df.head()

# nltk.download('wordnet')
# nltk.download('stopwords')
# nltk.download('punkt')
stop_words = nltk.corpus.stopwords.words(['english'])

print(stop_words)

lem = WordNetLemmatizer()

def cleaning(data):
    tweet_without_url = re.sub(r'http\S+',' ', data)
    tweet_without_hashtag = re.sub(r'#\w+', ' ', tweet_without_url)
    tweet_without_mentions = re.sub(r'@\w+',' ', tweet_without_hashtag)
    precleaned_tweet = re.sub('[^A-Za-z]+', ' ', tweet_without_mentions)

    # Tokenization
    tweet_tokens = TweetTokenizer().tokenize(precleaned_tweet)
    tokens_without_punc = [w for w in tweet_tokens if w.isalpha()]
    tokens_without_sw = [t for t in tokens_without_punc if t not in stop_words]

    text_cleaned = [lem.lemmatize(t) for t in tokens_without_sw]

    return " ".join(text_cleaned)

df['cleaned_tweets'] = df['tweets'].apply(lambda x: cleaning(x))
df['date'] = df_clean['date']
df['date_clean'] = pd.to_datetime(df['date']).dt.strftime('%Y-%m-%d')
df.drop(columns='date',inplace=True)
df.head()

def get_Subjectivity(tweet):
    return TextBlob(tweet).sentiment.subjectivity

def get_Polarity(tweet):
    return TextBlob(tweet).sentiment.polarity

btc_usd_grouped = btc_usd_grouped.set_index('Date')

btc_usd_grouped

def btc_price_cate(score):
    if score < 1:
        return 'negative'
    elif score == 1:
        return 'neutral'
    else:
        return 'positive'


def observe_period(period):
    res = btc_usd_grouped['Adj Close'].shift(period)/btc_usd_grouped['Adj Close']
    res = res.apply(btc_price_cate)
    return res

# Explain in report
# 7 is the days bitcoin price if shifted, to we will predict for week returns
time_sentiment = observe_period(7)
time_sentiment.index

# CHECK
df['crypto_sentiment'] = df.date_clean.apply(lambda x: time_sentiment[x] if x in time_sentiment else np.nan)

# We are going to drop the days for which we cannot measure the crypto_sentiment with the shift
df_2 = df.copy()
df_2 = df_2.dropna()
df_2.shape, df.shape

df_2['subjectivity'] = df_2['cleaned_tweets'].apply(get_Subjectivity)
df_2['polarity'] = df_2['cleaned_tweets'].apply(get_Polarity)
print(df_2.shape)
df_2.head()

def getSentiment(score, threshold=0.5):
    if score < 0:
        return 'negative'
    elif 0.001 < score < threshold:
        return 'neutral'
    else:
        return 'positive'

df_2['sentiment'] = df_2['polarity'].apply(getSentiment)
df_2['target'] = df_2['sentiment'] == df_2['crypto_sentiment']
df_2.head()

df_2.target.value_counts(normalize=True)

df_2.crypto_sentiment.value_counts()

df_2.sentiment.value_counts(normalize=True)

df_2 = df_2.set_index('date_clean')
df_2.head()

df_2.to_csv(f'{directory}/data/sampled_data/tweets_nlp_modelling_v3.csv')
