==> code/bitcoin/06_tweets_bow.ipynb <==
## Data processing for BOW

This notebooks receives the cleaned tweets from the file tweets_nlp_modelling_V2. Which is a second version with stratified sampling.

This file will be processed to obtain the tokens and then apply the BOW function.
Tokens will be grouped by date, in order to include all the tokens for each day.

Libraries and stopwords:

import pandas as pd
from itertools import chain
import nltk
import joblib
from functions.tweets_tokenization import \
    tokenize_tweets, \
    dictionary_tweets, \
    bow_tweets

stopwords = nltk.corpus.stopwords.words(['english'])

Import the data from CSV:

Import the bitcoin and tweets files.

directory = '~/PycharmProjects/tfm_hugopobil'
df = pd.read_csv(f'{directory}/data/sampled_data/tweets_nlp_modelling_v3.csv')
btc_usd_grouped = pd.read_csv(f'{directory}/data/sampled_data/btc_usd_grouped_v2.csv')
df = df.set_index('date_clean')

print(df.shape)
df.head()

### Tweets tokenization without grouping

### Group TOKENS by date:

This will create a dataframe with the accumulation of all tokens with the same date to obtain a grouped tokens.

df['tokens'] = tokenize_tweets(df.tweets.to_list())

# df_grouped = df.groupby(df.index).agg({'tokens': lambda x: list(chain(*x.to_list()))})

### Model Data Preparation

X = BOW for grouped tweets by days
Y (target) = Bitcoin returns

df

Join both dataframes and drop NA values, so we end up with a dataframe that includes the tokens and daily returns as our target variable for Tweets and Bitcoin.

Calculate BOW and Train Set for documents with less than 3 words.

dictionary_model_data = dictionary_tweets(df['tokens'])
X_model_data, doc2bow_model_data = bow_tweets(df['tokens'], dictionary_model_data)
X_model_data.shape

Save to local:

joblib.dump(dictionary_model_data, '/Users/hpp/PycharmProjects/tfm_hugopobil/models/topic_analisis/dictionary.joblib')

joblib.dump(doc2bow_model_data, '/Users/hpp/PycharmProjects/tfm_hugopobil/models/topic_analisis/doc2bow.joblib')
df.to_csv('/Users/hpp/PycharmProjects/tfm_hugopobil//models/topic_analisis/model_data.csv')
joblib.dump(X_model_data, '/Users/hpp/PycharmProjects/tfm_hugopobil/models/topic_analisis/X_model_data.joblib')
